refactor: isolate HTTP error converters
